Fixed bug where a playlist called "Swinging Safari" would cause a playlist called "Swing" to also be highlighted in tree.
Fixed a more serious issue where if a file is missing then its tags would be wiped out.
Now it will add the filename to the list of filenames so that the tags are still saved.
And it will add the filename to the track map with a specially constructed track that is flagged as "missing".
A missing track is therefore stil displayed in the table, but it will:
- Have a grey color
- The checkbox cannot be selected
- The popup menu won't allow the track to be played or chosen as next tracl/cortina/tanda.